+++ angular-workspace/components/toggletip/toggletip.stories.ts
@@ -14,7 +14,7 @@ const meta: Meta<ToggletipArgs> = {
       imports: [DsoToggletip],
     }),
   ],
-  title: "Toggletip",
+  title: "Toggletip (Deprecated)",
 };
 
 export default meta;