+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Added
+* Typografie Headings: Aanpassing kleur in context van Accordion compact ([#3561](https://github.com/dso-toolkit/dso-toolkit/issues/3561))
+
 ### Changed
 * History Item: Current state toevoegen + kopkleur aanpassen en chevron ([#3423](https://github.com/dso-toolkit/dso-toolkit/issues/3423))
 * List: Variant unordered-action toevoegen ([#3511](https://github.com/dso-toolkit/dso-toolkit/issues/3511))